added early timing task

+++ flight-software/tasks/obdh.py
@@ -12,18 +12,19 @@ class Task(DebugTask):
     ID = 0x02
 
     # Class variables
+    SD_cleaned = False # TODO DEBUG
     SD_scanned = False
     SD_stored_volume = 0
 
     async def main_task(self):
 
         if SM.current_state == "STARTUP":
-            DH.delete_all_files()
+            if self.SD_cleaned == False:
+                DH.delete_all_files()
+                self.SD_cleaned = True
             if self.SD_scanned == False:
                 DH.scan_SD_card()
                 self.SD_scanned = True
-            # TODO Temporarily start global state switch here
-            SM.switch_to("NOMINAL")
         elif SM.current_state == "NOMINAL":
             # DH.clean_up()
             self.SD_stored_volume = DH.compute_total_size_files()
